dom/workers/test/test_promise.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Promise object in workers</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

  function runTest() {
    var worker = new Worker("promise_worker.js");

    worker.onmessage = function(event) {

      if (event.data.type == 'finish') {
        SimpleTest.finish();
      } else if (event.data.type == 'status') {
        ok(event.data.status, event.data.msg);
      }
    }

    worker.onerror = function(event) {
      ok(false, "Worker had an error: " + event.data);
      SimpleTest.finish();
    };

    worker.postMessage(true);
  }

  SimpleTest.waitForExplicitFinish();
  runTest();
</script>
</pre>
</body>
</html>
